Struct ferrite::BufferDesc

pub struct BufferDesc { /* fields omitted */ }

Builder structure specifying the parameters of a newly created buffer object

Methods

impl BufferDesc

fn new(byte_size: usize, usage: BufferUsage) -> Self

fn sharing_queue_families(&mut self, indices: Vec<u32>) -> &mut Self

A list of queue families that will access this buffer

fn sparse_binding_opt(&mut self, opt: BufferSparseBinding) -> &mut Self

A bitmask of BufferSparseBinding specifying additional parameters of the buffer

fn create(&self, device: &Device) -> Result<Buffer>

Create a new buffer object

Failure

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY