Struct ferrite::BufferUsage

pub struct BufferUsage(pub VkBufferUsageFlags);

Bitmask specifying allowed usage of a buffer

Methods

impl BufferUsage

const TRANSFER_SRC: Self

TRANSFER_SRC: Self = BufferUsage(VK_BUFFER_USAGE_TRANSFER_SRC_BIT)

Specifies that the buffer can be used as the source of a transfer command

const TRANSFER_DEST: Self

TRANSFER_DEST: Self = BufferUsage(VK_BUFFER_USAGE_TRANSFER_DST_BIT)

Specifies that the buffer can be used as the destination of a transfer command

const UNIFORM_TEXEL_BUFFER: Self

UNIFORM_TEXEL_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_UNIFORM_TEXEL_BUFFER_BIT)

Specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER

const STORAGE_TEXEL_BUFFER: Self

STORAGE_TEXEL_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT)

Specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER

const UNIFORM_BUFFER: Self

UNIFORM_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT)

Specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER or VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC

const STORAGE_BUFFER: Self

STORAGE_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_STORAGE_BUFFER_BIT)

Specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

const INDEX_BUFFER: Self

INDEX_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_INDEX_BUFFER_BIT)

Specifies that the buffer is suitable for passing as the buffer parameter to DrawCommandBuffer::bind_index_buffer

const VERTEX_BUFFER: Self

VERTEX_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_VERTEX_BUFFER_BIT)

Specifies that the buffer is suitable for passing as an element of the buffers array to DrawCommandBuffer::bind_vertex_buffers

const INDIRECT_BUFFER: Self

INDIRECT_BUFFER: Self = BufferUsage(VK_BUFFER_USAGE_INDIRECT_BUFFER_BIT)

Specifies that the buffer is suitable for passing as the buffer parameter to DrawCommandBuffer::draw_indirect, DrawCommandBuffer::draw_indexed_indirect, or ComputeCommandBuffer::dispatch_indirect

fn transfer_src(&self) -> Self

Specifies that the buffer can be used as the source of a transfer command

fn transfer_dest(&self) -> Self

Specifies that the buffer can be used as the destination of a transfer command

fn uniform_texel_buffer(&self) -> Self

Specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type VK_DESCRIPTOR_TYPE_UNIFORM_TEXEL_BUFFER

fn storage_texel_buffer(&self) -> Self

Specifies that the buffer can be used to create a BufferView suitable for occupying a DescriptorSet slot of type VK_DESCRIPTOR_TYPE_STORAGE_TEXEL_BUFFER

fn uniform_buffer(&self) -> Self

Specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER or VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC

fn storage_buffer(&self) -> Self

Specifies that the buffer can be used in a DescriptorBufferInfo suitable for occupying a DescriptorSet slot either of type VK_DESCRIPTOR_TYPE_STORAGE_BUFFER or VK_DESCRIPTOR_TYPE_STORAGE_BUFFER_DYNAMIC

fn index_buffer(&self) -> Self

Specifies that the buffer is suitable for passing as the buffer parameter to DrawCommandBuffer::bind_index_buffer

fn vertex_buffer(&self) -> Self

Specifies that the buffer is suitable for passing as an element of the buffers array to DrawCommandBuffer::bind_vertex_buffers

fn indirect_buffer(&self) -> Self

Specifies that the buffer is suitable for passing as the buffer parameter to DrawCommandBuffer::draw_indirect, DrawCommandBuffer::draw_indexed_indirect, or ComputeCommandBuffer::dispatch_indirect