Struct ferrite::CmdRecord

pub struct CmdRecord<'d> { /* fields omitted */ }

The recording state of commandbuffers

Methods

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics/Compute Commands: Pipeline Setup

fn bind_graphics_pipeline(
&mut self,
pipeline: &Pipeline,
layout: &'d PipelineLayout
) -> &mut Self

Bind a pipeline object to a command buffer

fn bind_compute_pipeline(
&mut self,
pipeline: &Pipeline,
layout: &'d PipelineLayout
) -> &mut Self

Bind a pipeline object to a command buffer

fn bind_graphics_descriptor_sets(
&mut self,
first: u32,
descriptor_sets: &[VkDescriptorSet],
dynamic_offsets: &[u32]
) -> &mut Self

Binds descriptor sets to a command buffer

fn bind_compute_descriptor_sets(
&mut self,
first: u32,
descriptor_sets: &[VkDescriptorSet],
dynamic_offsets: &[u32]
) -> &mut Self

Binds descriptor sets to a command buffer

fn push_graphics_constants<T>(
&mut self,
stage: ShaderStage,
offset: u32,
values: &[T]
) -> &mut Self

Update the values of push constants

fn push_compute_constants<T>(
&mut self,
stage: ShaderStage,
offset: u32,
values: &[T]
) -> &mut Self

Update the values of push constants

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Updating dynamic states

fn set_viewport(&mut self, first: u32, viewports: &[VkViewport]) -> &mut Self

Set the viewport on a command buffer

fn set_scissor(&mut self, first: u32, scissors: &[VkRect2D]) -> &mut Self

Set the dynamic scissor rectangles on a command buffer

fn set_line_width(&mut self, w: f32) -> &Self

Set the dynamic line width state

fn set_depth_bias(
&mut self,
constant_factor: f32,
clamp: f32,
slope_factor: f32
) -> &mut Self

Set the depth bias dynamic state

fn set_blend_constants(&mut self, blend_constants: [f32; 4]) -> &mut Self

Set the values of blend constants

fn set_depth_bounds(&mut self, bounds: Range<f32>) -> &mut Self

Set the depth bounds test values for a command buffer

fn set_stencil_compare_mask(
&mut self,
face_mask: StencilFaceMask,
compare_mask: u32
) -> &mut Self

Set the stencil compare mask dynamic state

fn set_stencil_write_mask(
&mut self,
face_mask: StencilFaceMask,
write_mask: u32
) -> &mut Self

Set the stencil write mask dynamic state

fn set_stencil_reference(
&mut self,
face_mask: StencilFaceMask,
reference: u32
) -> &mut Self

Set the stencil reference dynamic state

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Binding Buffers

fn bind_index_buffer(
&mut self,
buffer: &Buffer,
offset: usize,
index_type: IndexType
) -> &mut Self

Bind an index buffer to a command buffer

fn bind_vertex_buffers(
&mut self,
first: u32,
buffers: &[(&Buffer, usize)]
) -> &mut Self

Bind vertex buffers to a command buffer

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Inside a Render Pass

fn draw(
&mut self,
vertex_count: u32,
instance_count: u32,
first_vertex: u32,
first_instance: u32
) -> &mut Self

Draw primitives

fn draw_indexed(
&mut self,
index_count: u32,
instance_count: u32,
first_index: u32,
vertex_offset: i32,
first_instance: u32
) -> &mut Self

Issue an indexed draw into a command buffer

fn draw_indirect(
&mut self,
buffer: &Buffer,
offset: usize,
draw_count: u32,
stride: u32
) -> &mut Self

Issue an indirect draw into a command buffer

fn draw_indexed_indirect(
&mut self,
buffer: &Buffer,
offset: usize,
draw_count: u32,
stride: u32
) -> &mut Self

Perform an indexed indirect draw

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Compute Commands: Dispatching kernels

fn dispatch(
&mut self,
group_count_x: u32,
group_count_y: u32,
group_count_z: u32
) -> &mut Self

Dispatch compute work items

fn dispatch_indirect(&mut self, buffer: &Buffer, offset: usize) -> &mut Self

Dispatch compute work items using indirect parameters

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Transfer Commands: Copying resources

fn copy_buffer(
&mut self,
src: &Buffer,
dst: &Buffer,
regions: &[VkBufferCopy]
) -> &mut Self

Copy data between buffer regions

fn copy_image(
&mut self,
src: &Image,
src_layout: ImageLayout,
dst: &Image,
dst_layout: ImageLayout,
regions: &[VkImageCopy]
) -> &mut Self

Copy data between images

fn blit_image(
&mut self,
src: &Image,
src_layout: ImageLayout,
dst: &Image,
dst_layout: ImageLayout,
regions: &[VkImageBlit],
filter: FilterMode
) -> &mut Self

Copy regions of an image, potentially performing format conversion

fn copy_buffer_to_image(
&mut self,
src_buffer: &Buffer,
dst_image: &Image,
dst_layout: ImageLayout,
regions: &[VkBufferImageCopy]
) -> &mut Self

Copy data from a buffer into an image

fn copy_image_to_buffer(
&mut self,
src_image: &Image,
src_layout: ImageLayout,
dst_buffer: &Buffer,
regions: &[VkBufferImageCopy]
) -> &mut Self

Copy image data into a buffer

fn update_buffer<T>(
&mut self,
dst: &Buffer,
dst_offset: usize,
size: usize,
data: &T
) -> &mut Self

Update a buffer's contents from host memory

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics/Compute Commands: Transfer-like(clearing/filling) commands

fn fill_buffer(
&mut self,
dst: &Buffer,
dst_offset: usize,
size: usize,
data: u32
) -> &mut Self

Fill a region of a buffer with a fixed value.
size is number of bytes to fill

fn clear_color_image<T: ClearColorValue>(
&mut self,
image: &Image,
layout: ImageLayout,
color: &T,
ranges: &[VkImageSubresourceRange]
) -> &mut Self

Clear regions of a color image

fn clear_depth_stencil_image(
&mut self,
image: &Image,
layout: ImageLayout,
depth: f32,
stencil: u32,
ranges: &[VkImageSubresourceRange]
) -> &mut Self

Fill regions of a combined depth/stencil image

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Attachment clearing

fn clear_attachments(
&mut self,
attachments: &[VkClearAttachment],
rects: &[VkClearRect]
) -> &mut Self

Clear regions within currently bound framebuffer attachments

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Resolving an image to another image

fn resolve_image(
&mut self,
src: &Image,
src_layout: ImageLayout,
dst: &Image,
dst_layout: ImageLayout,
regions: &[VkImageResolve]
) -> &mut Self

Resolve regions of an image

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics/Compute Commands: Synchronization between command buffers/queues

fn set_event(
&mut self,
event: &Event,
stage_mask: PipelineStageFlags
) -> &mut Self

Set an event object to signaled state

fn reset_event(
&mut self,
event: &Event,
stage_mask: PipelineStageFlags
) -> &mut Self

Reset an event object to non-signaled state

fn wait_events(
&mut self,
events: &[&Event],
src_stage_mask: PipelineStageFlags,
dst_stage_mask: PipelineStageFlags,
memory_barriers: &[VkMemoryBarrier],
buffer_memory_barriers: &[VkBufferMemoryBarrier],
image_memory_barriers: &[VkImageMemoryBarrier]
) -> &mut Self

Wait for one or more events and insert a set of memory

fn pipeline_barrier(
&mut self,
src_stage_mask: PipelineStageFlags,
dst_stage_mask: PipelineStageFlags,
by_region: bool,
memory_barriers: &[VkMemoryBarrier],
buffer_memory_barriers: &[VkBufferMemoryBarrier],
image_memory_barriers: &[VkImageMemoryBarrier]
) -> &mut Self

Insert a memory dependency

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics/Compute Commands: Querying

fn begin_query(
&mut self,
pool: &QueryPool,
query: u32,
precise_query: bool
) -> &mut Self

Begin a query

fn end_query(&mut self, pool: &QueryPool, query: u32) -> &mut Self

Ends a query

fn reset_query_pool(&mut self, pool: &QueryPool, range: Range<u32>) -> &mut Self

Reset queries in a query pool

fn write_timestamp(
&mut self,
stage: PipelineStageFlags,
pool: &QueryPool,
query: u32
) -> &mut Self

Write a device timestamp into a query object

fn copy_query_pool_results(
&mut self,
pool: &QueryPool,
range: Range<u32>,
dst: &Buffer,
dst_offset: usize,
stride: usize,
wide_result: bool,
flags: QueryResultFlags
) -> &mut Self

Copy the results of queries in a query pool to a buffer object

impl<'d> CmdRecord<'d>

[feature = "FeImplements"] Graphics Commands: Manipulating with Render Passes

fn begin_render_pass(
&mut self,
pass: &RenderPass,
framebuffer: &Framebuffer,
render_area: VkRect2D,
clear_values: &[ClearValue],
inline_commands: bool
) -> &mut Self

Begin a new render pass

fn next_subpass(&mut self, inline_commands: bool) -> &mut Self

Transition to the next subpass of a render pass

fn end_render_pass(&mut self) -> &mut Self

End the current render pass