Struct ferrite::DSLBindings

pub struct DSLBindings { /* fields omitted */ }

Structure specifying a descriptor set layout binding

// Create bindings for 1 uniform buffer (at binding #0) and 2 combined image samplers (at binding #1)
let bindings = DSLBindings::new().uniform_buffers(1, ShaderStage::VERTEX)
  .combined_image_samplers(2, ShaderStage::FRAGMENT);

Methods

impl DSLBindings

pub fn new() -> Self

pub fn uniform_buffers(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

The elements of the VkWriteDescriptorSet::pBufferInfo array of VkDescriptorBufferInfo structures will be used to update the descriptors, and other arrays will be ignored

Panics

Calling uniform_buffer twice or more is invalid due to Vulkan restriction

pub fn storage_buffers(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn uniform_buffers_dynamic(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn storage_buffers_dynamic(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn uniform_texel_buffers(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn storage_texel_buffers(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn samplers(
&mut self,
count: u32,
shader_visibility: ShaderStage,
imm_samplers: Vec<&Sampler>
) -> &mut Self

pub fn combined_image_samplers(
&mut self,
count: u32,
shader_visibility: ShaderStage,
imm_samplers: Vec<&Sampler>
) -> &mut Self

pub fn sampled_images(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self

pub fn storage_images(
&mut self,
count: u32,
shader_visibility: ShaderStage
) -> &mut Self