Struct ferrite::DebugReportCallbackBuilder

pub struct DebugReportCallbackBuilder<'i> { /* fields omitted */ }

Methods

impl<'i> DebugReportCallbackBuilder<'i>

pub fn new(
instance: &'i Instance,
callback: PFN_vkDebugReportCallbackEXT
) -> Self

Create a builder object of DebugReportCallbackBuilder from instance, called back to callback

pub fn report_error(&mut self) -> &mut Self

Reports an error that may cause undefined results, including an application crash

pub fn report_warning(&mut self) -> &mut Self

Reports an unexpected use. e.g. Not destroying objects prior to destroying the containing object or potential inconsistencies between descriptor set layout and the layout in the corresponding shader, etc

pub fn report_performance_warning(&mut self) -> &mut Self

Reports a potentially non-optimal use of Vulkan. e.g. using vkCmdClearColorImage when a RenderPass load_op would have worked

pub fn report_information(&mut self) -> &mut Self

Reports an informational message such as resource details that may be handy when debugging an application

pub fn report_debug_information(&mut self) -> &mut Self

Reports diagnostic information from the loader and layers

pub fn create(&mut self) -> Result<DebugReportCallback>

Register a debug report callback

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY