Struct ferrite::Device

pub struct Device(_);

Opaque handle to a device object

Methods

impl Device

Following methods are enabled with [feature = "FeImplements"]

pub fn extra_procedure<F: FnTransmute>(&self, name: &str) -> Option<F>

Return a function pointer for a command

Failures

If function is not provided by instance or name is empty, returns None

pub fn queue(&self, family_index: u32, queue_index: u32) -> Queue

Get a queue handle from a device

pub fn flush_memory_range(&self, ranges: Vec<MappedMemoryRange>) -> Result<()>

Flush MappedMemoryRanges Flushing the memory range allows that host writes to the memory ranges can be made available to device access

pub fn invalidate_memory_range(
&self,
ranges: Vec<MappedMemoryRange>
) -> Result<()>

Invalidate MappedMemoryRanges Invalidating the memory range allows that device writes to the memory ranges which have been made visible to the VK_ACCESS_HOST_WRITE_BIT and VK_ACCESS_HOST_READ_BIT are made visible to the host

pub fn update_descriptor_sets(
&self,
write: &[DescriptorSetWriteInfo],
copy: &[DescriptorSetCopyInfo]
)

Update the contents of a descriptor set object

impl Device

Following methods are enabled with [feature = "FeImplements"]

pub fn create_graphics_pipelines(
&self,
builders: &[GraphicsPipelineBuilder],
cache: Option<&PipelineCache>
) -> Result<Vec<Pipeline>>

Create graphics pipelines

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY