Struct ferrite::Fence

pub struct Fence(pub VkFence, _);

Opaque handle to a fence object

Methods

impl Fence

fn new(device: &Device, signaled: bool) -> Result<Self>

Create a new fence object

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

impl Fence

fn wait_multiple(
objects: &[&Self],
wait_all: bool,
timeout: Option<u64>
) -> Result<bool>

Wait for one or more fences to become signaled, returns Ok(true) if operation is timed out

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

fn wait_timeout(&self, timeout: u64) -> Result<bool>

Wait for a fence to become signaled, returns Ok(true) if operation is timed out

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

fn reset_multiple(objects: &[&Self]) -> Result<()>

Resets one or more fence objects

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn reset(&self) -> Result<()>

Resets a fence object

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY