Struct ferrite::GraphicsPipelineBuilder

pub struct GraphicsPipelineBuilder<'d> { /* fields omitted */ }

Builder struct to construct a Pipeline for graphics operations

Methods

impl<'d> GraphicsPipelineBuilder<'d>

fn new(layout: &'d PipelineLayout, rpsp: (&'d RenderPass, u32)) -> Self

Initialize the builder object

impl<'d> GraphicsPipelineBuilder<'d>

Shading State and Input Configuration

fn vertex_shader(&mut self, shader: PipelineShader<'d>) -> &mut Self

Set the vertex shader in this pipeline

fn tessellation_control_shader(
&mut self,
shader: PipelineShader<'d>
) -> &mut Self

Set the tessellation control shader(hull shader) in this pipeline

fn tessellation_evaluation_shader(
&mut self,
shader: PipelineShader<'d>
) -> &mut Self

Set the tessellation evaluation shader(domain shader) in this pipeline

fn geometry_shader(&mut self, shader: PipelineShader<'d>) -> &mut Self

Set the geometry shader in this pipeline

fn fragment_shader(&mut self, shader: PipelineShader<'d>) -> &mut Self

Set the fragment shader in this pipeline

fn vertex_input_state(
&mut self,
bindings: Vec<VkVertexInputBindingDescription>,
attributes: Vec<VkVertexInputAttributeDescription>
) -> &mut Self

Set the vertex input layout in this pipeline

fn vertex_processing(
&mut self,
shader: PipelineShader<'d>,
bindings: Vec<VkVertexInputBindingDescription>,
attributes: Vec<VkVertexInputAttributeDescription>
) -> &mut Self

Set the vertex processing state(a shader and an input layout) in this pipeline

fn primitive_topology(
&mut self,
topo: VkPrimitiveTopology,
enable_restarting: bool
) -> &mut Self

The primitive topology and primitiveRestartEnable: controls whether a special vertex index value is treated as restarting the assembly of primitives.
The special index value is either

0xFFFFFFFF when the index type is equal to VK_INDEX_TYPE_UINT32, or

0xFFFF when the index type is equal to VK_INDEX_TYPE_UINT16.

Primitive restart is not allowed for "list" topologies

fn patch_control_point_count(&mut self, count: u32) -> &mut Self

Number of control points per patch

fn tessellator_settings(
&mut self,
control: PipelineShader<'d>,
evaluation: PipelineShader<'d>,
num_control_points: u32
) -> &mut Self

Set the tessellation processing state(hull/domain shaders and a number of control points)

impl<'d> GraphicsPipelineBuilder<'d>

Viewport / Scissor State

unsafe fn viewports(&mut self, vps: DynamicArrayState<VkViewport>) -> &mut Self

Safety

Application must guarantee that the number of viewports and scissors are identical

unsafe fn scissors(&mut self, scs: DynamicArrayState<VkRect2D>) -> &mut Self

Safety

Application must guarantee that the number of viewports and scissors are identical

fn fixed_viewport_scissors(
&mut self,
vps: DynamicArrayState<VkViewport>,
scissor: DynamicArrayState<VkRect2D>
) -> &mut Self

Safety way calling viewports and scissors

impl<'d> GraphicsPipelineBuilder<'d>

Rasterization State

fn depth_clamp_enable(&mut self, enable: bool) -> &mut Self

Controls whether to clamp the fragment's depth values instead of clipping primitives to the z planes of the frustum, as described in Primitive Clipping in Vulkan Specification

fn rasterizer_discard_enable(&mut self, enable: bool) -> &mut Self

Controls whether primitives are discarded immediately before the rasterization stage

fn polygon_mode(&mut self, mode: VkPolygonMode) -> &mut Self

The triangle rendering mode

fn cull_mode(&mut self, mode: VkCullModeFlags) -> &mut Self

The triangle facing direction used for primitive culling

fn front_face(&mut self, face: VkFrontFace) -> &mut Self

The front-facing triangle orientation to be used for culling

fn depth_bias(
&mut self,
opts: SwitchOrDynamicState<(f32, f32, f32)>
) -> &mut Self

Specify None to disable to bias fragment depth values.
Tuple Member: (ConstantFactor, Clamp, SlopeFactor)

ConstantFactor: A scalar factor controlling the constant depth value added to each fragment

Clamp: The maximum (or minimum) depth bias of a fragment

SlopeFactor: A scalar factor applied to a fragment's slope in depth bias calculations

fn line_width(&mut self, width: Option<f32>) -> &mut Self

The width of rasterized line segments. Specifying None means that the lineWidth parameter is a dynamic state.

impl<'d> GraphicsPipelineBuilder<'d>

Multisample State

fn shading_per_sample(&mut self, flag: bool) -> &mut Self

Fragment shading executed per-sample(true), or per-fragment(false), as described in Sample Shading in Vulkan Specification

fn min_sample_shading(&mut self, frac: f32) -> &mut Self

The minimum fraction of sample shading, as described in Sample Shading in Vulkan Specification

fn alpha_to_coverage_enable(&mut self, enable: bool) -> &mut Self

Controls whether a temporary coverage value is generated based on the alpha component of the fragment's first color output as specified in the Multisample Coverage section in Vulkan Specification

fn alpha_to_one_enable(&mut self, enable: bool) -> &mut Self

Controls whether the alpha component of the fragment's first color output is replaced with one as described in Multisample Coverage section in Vulkan Specification

fn rasterization_samples(
&mut self,
bits: u32,
sample_masks: Vec<VkSampleMask>
) -> &mut Self

A sample count bits specifying the number of samples per pixel used in rasterization

impl<'d> GraphicsPipelineBuilder<'d>

Depth/Stencil State

fn clear_depth_stencil_state(&mut self) -> &mut Self

Clear depth/stencil state

fn depth_test_enable(&mut self, enable: bool) -> &mut Self

Controls whether depth testing is enabled

fn depth_write_enable(&mut self, enable: bool) -> &mut Self

Controls whether depth writes are enabled, or always disabled

fn depth_compare_op(&mut self, op: CompareOp) -> &mut Self

The comparison operator used in the depth test

fn depth_test_settings(
&mut self,
compare_op: Option<CompareOp>,
write_enable: bool
) -> &mut Self

Controls whether depth testing is enabled, depth writes are enabled, and the comparison operator used in the depth test Specifying None to compare_to disables depth testing

fn depth_bounds_test_enable(&mut self, enable: bool) -> &mut Self

Controls whether depth bounds testing is enabled

fn stencil_test_enable(&mut self, enable: bool) -> &mut Self

Controls whether stencil testing is enabled

fn stencil_control_front(
&mut self,
compare_op: CompareOp,
reference: u32,
compare_mask: u32,
fail_op: StencilOp,
pass_op: StencilOp,
depth_fail_op: StencilOp,
write_mask: u32
) -> &mut Self

Control the parameter of the stencil test

fn stencil_control_back(
&mut self,
compare_op: CompareOp,
reference: u32,
compare_mask: u32,
fail_op: StencilOp,
pass_op: StencilOp,
depth_fail_op: StencilOp,
write_mask: u32
) -> &mut Self

Control the parameter of the stencil test

fn stencil_compare_mask(&mut self, mask: Option<(u32, u32)>) -> &mut Self

Controls the parameter of the compare mask of the stencil test. Tuple ordering: (front, back). Specifying None means that the parameter is a dynamic state

fn stencil_write_mask(&mut self, mask: Option<(u32, u32)>) -> &mut Self

Controls the parameter of the write mask of the stencil test. Tuple ordering: (front, back) Specifying None means that the parameter is a dynamic state

fn stencil_reference(&mut self, mask: Option<(u32, u32)>) -> &mut Self

Controls the parameter of the reference of the stencil test. Tuple ordering: (front, back) Specifying None means that the parameter is a dynamic state

fn depth_bounds_range(&mut self, bounds: Range<f32>) -> &mut Self

The range of values used in the depth bounds test

fn depth_bounds(
&mut self,
bounds: SwitchOrDynamicState<Range<f32>>
) -> &mut Self

Control the depth bounds test

impl<'d> GraphicsPipelineBuilder<'d>

Color Blending

fn logic_op(&mut self, op: Option<LogicOp>) -> &mut Self

Which logical operation to apply. Specifying None disables Logical Operations

fn add_attachment_blend(
&mut self,
blend: VkPipelineColorBlendAttachmentState
) -> &mut Self

Per target attachment states

fn blend_constants(&mut self, values: Option<[f32; 4]>) -> &mut Self

Array of four values used as the R, G, B, and A components of the blend constant that are used in blending, depending on the blend factor. Specifying None means that the blendConstants parameter is a dynamic state

impl<'d> GraphicsPipelineBuilder<'d>

Misc Configurations

fn base(&mut self, b: BasePipeline<'d>) -> &mut Self

The base pipeline handle/index to derive from

fn layout(&mut self, l: &'d PipelineLayout) -> &mut Self

The description of binding locations used by both the pipeline and descriptor sets used with the pipeline

fn render_pass(&mut self, rpo: &'d RenderPass, subpass: u32) -> &mut Self

A handle to a render pass object and the index of the subpass where this pipeline will be used

fn enable_optimization(&mut self, opt: bool) -> &mut Self

The created pipeline will or will not be optimized.
Disabling optimization of the pipeline may reduce the time taken to create the pipeline

fn allow_derivatives(&mut self, allow: bool) -> &mut Self

The pipeline to be created is allowed to be the parent of a pipeline that will be created in a subsequent creation operation

impl<'d> GraphicsPipelineBuilder<'d>

Unsafe Utilities

unsafe fn vertex_input_state_create_info(
&mut self,
state: VkPipelineVertexInputStateCreateInfo
) -> &mut Self

Set the VkPipelineVertexInputStateCreateInfo structure directly

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn input_assembly_state_create_info(
&mut self,
state: VkPipelineInputAssemblyStateCreateInfo
) -> &mut Self

Set the VkPipelineInputAssemblyStateCreateInfo structure directly

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn tessellation_state_create_info(
&mut self,
state: Option<Box<VkPipelineTessellationStateCreateInfo>>
) -> &mut Self

Set the VkPipelineTessellationStateCreateInfo structure directly

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn viewport_state_create_info(
&mut self,
state: Option<Box<VkPipelineViewportStateCreateInfo>>
) -> &mut Self

Set the VkPipelineViewportStateCreateInfo structure directly. This does not clear any dynamic states

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn rasterization_state_create_info(
&mut self,
state: VkPipelineRasterizationStateCreateInfo
) -> &mut Self

Set the VkPipelineRasterizationStateCreateInfo structure directly. This does not clear any dynamic states

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn multisample_state_create_info(
&mut self,
state: Option<Box<VkPipelineMultisampleStateCreateInfo>>
) -> &mut Self

Set the VkPipelineMultisampleStateCreateInfo structure directly. This does not clear any dynamic states

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn depth_stencil_state_create_info(
&mut self,
state: Option<Box<VkPipelineDepthStencilStateCreateInfo>>
) -> &mut Self

Set the VkPipelineDepthStencilStateCreateInfo structure directly. This does not clear any dynamic states

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid

unsafe fn color_blend_state_info(
&mut self,
state: Option<Box<VkPipelineColorBlendStateCreateInfo>>
) -> &mut Self

Set the VkPipelineColorBlendStateCreateInfo structure directly. This does not clear any dynamic states

Safety

Application must guarantee these constraints:

The lifetime of the content in the structure is valid for this builder

The content in the structure is valid