Struct ferrite::ImageFlags

pub struct ImageFlags(pub VkImageCreateFlags);

Bitmask specifying additional parameters of an image

Methods

impl ImageFlags

const EMPTY: Self

EMPTY: Self = ImageFlags(0)

Empty bits

const SPARSE_BINDING: Self

SPARSE_BINDING: Self = ImageFlags(VK_IMAGE_CREATE_SPARSE_BINDING_BIT)

The image will be backed using sparse memory binding

const SPARSE_RESIDENCY: Self

SPARSE_RESIDENCY: Self = ImageFlags(VK_IMAGE_CREATE_SPARSE_BINDING_BIT |
               VK_IMAGE_CREATE_SPARSE_RESIDENCY_BIT)

The image can be partially backed using sparse memory binding. This bit is with SPARSE_BINDING implicitly

const SPARSE_ALIASED: Self

SPARSE_ALIASED: Self = ImageFlags(VK_IMAGE_CREATE_SPARSE_BINDING_BIT |
               VK_IMAGE_CREATE_SPARSE_ALIASED_BIT)

The image will be backed using sparse memory binding with memory ranges that might also simultaneously be backing another image. This bit is with SPARSE_BINDING implicitly

const MUTABLE_FORMAT: Self

MUTABLE_FORMAT: Self = ImageFlags(VK_IMAGE_CREATE_MUTABLE_FORMAT_BIT)

The image can be used to create a ImageView with a different format from the image

const CUBE_COMPATIBLE: Self

CUBE_COMPATIBLE: Self = ImageFlags(VK_IMAGE_CREATE_CUBE_COMPATIBLE_BIT)

The image can be used to create a ImageView of type ImageViewType::Cube or ImageViewType::CubeArray

fn sparse_binding(&self) -> Self

The image will be backed using sparse memory binding

fn sparse_residency(&self) -> Self

The image can be partially backed using sparse memory binding. This bit is with SPARSE_BINDING implicitly

fn sparse_aliased(&self) -> Self

The image will be backed using sparse memory binding with memory ranges that might also simultaneously be backing another image. This bit is with SPARSE_BINDING implicitly

fn mutable_format(&self) -> Self

The image can be used to create a ImageView with a different format from the image

fn cube_compatible(&self) -> Self

The image can be used to create a ImageView of type ImageViewType::Cube or ImageViewType::CubeArray