Struct ferrite::ImageUsage

pub struct ImageUsage(pub VkImageUsageFlags);

Bitmask specifying intended usage of an image

Methods

impl ImageUsage

const TRANSFER_SRC: Self

TRANSFER_SRC: Self = ImageUsage(VK_IMAGE_USAGE_TRANSFER_SRC_BIT)

The image can be used as the source of a transfer command

const TRANSFER_DEST: Self

TRANSFER_DEST: Self = ImageUsage(VK_IMAGE_USAGE_TRANSFER_DST_BIT)

The image can be used as the destination of a transfer command

const SAMPLED: Self

SAMPLED: Self = ImageUsage(VK_IMAGE_USAGE_SAMPLED_BIT)

The image can be used to create ImageView suitable for occupying a DescriptorSet slot either of type DescriptorType::SampledImage or DescriptorType::CombinedImageSampler, and be sampled by a shader

const STORAGE: Self

STORAGE: Self = ImageUsage(VK_IMAGE_USAGE_STORAGE_BIT)

The image can be used to create a ImageView suitable for occupying a DescriptorSet slot of type DescriptorType::StorageImage

const COLOR_ATTACHMENT: Self

COLOR_ATTACHMENT: Self = ImageUsage(VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT)

The image can be used to create a ImageView suitable for use as a color or resolve attachment in a Framebuffer

const DEPTH_STENCIL_ATTACHMENT: Self

DEPTH_STENCIL_ATTACHMENT: Self = ImageUsage(VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT)

The image can be used to create a ImageView suitable for use as a depth/stencil attachment in a Framebuffer

const TRANSIENT_ATTACHMENT: Self

TRANSIENT_ATTACHMENT: Self = ImageUsage(VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT)

The memory bound to this image will have been allocated with the VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT This bit can be set for any image that can be used to create a ImageView suitable for use as a color, resolve, depth/stencil, or input attachment

const INPUT_ATTACHMENT: Self

INPUT_ATTACHMENT: Self = ImageUsage(VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT)

The image can be used to create a ImageView suitable for occupying DescriptorSet slot of type DescriptorType::InputAttachment; be read from a shader as an input attachment; and be used as an input attachment in a framebuffer

fn transfer_src(&self) -> Self

The image can be used as the source of a transfer command

fn transfer_dest(&self) -> Self

The image can be used as the destination of a transfer command

fn sampled(&self) -> Self

The image can be used to create ImageView suitable for occupying a DescriptorSet slot either of type DescriptorType::SampledImage or DescriptorType::CombinedImageSampler, and be sampled by a shader

fn storage(&self) -> Self

The image can be used to create a ImageView suitable for occupying a DescriptorSet slot of type DescriptorType::StorageImage

fn color_attachment(&self) -> Self

The image can be used to create a ImageView suitable for use as a color or resolve attachment in a Framebuffer

fn depth_stencil_attachment(&self) -> Self

The image can be used to create a ImageView suitable for use as a depth/stencil attachment in a Framebuffer

fn transient_attachment(&self) -> Self

The memory bound to this image will have been allocated with the VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT This bit can be set for any image that can be used to create a ImageView suitable for use as a color, resolve, depth/stencil, or input attachment

fn input_attachment(&self) -> Self

The image can be used to create a ImageView suitable for occupying DescriptorSet slot of type DescriptorType::InputAttachment; be read from a shader as an input attachment; and be used as an input attachment in a framebuffer