Struct ferrite::Instance

pub struct Instance(_);

Opaque handle to a instance object

Methods

impl Instance

Following methods are enabled with [feature = "FeImplements"]

fn extra_procedure<F: FnTransmute>(&self, name: &str) -> Option<F>

Return a function pointer for a command

Failures

If function is not provided by instance or name is empty, returns None

fn enumerate_physical_devices(&self) -> Result<Vec<PhysicalDevice>>

Enumerates the physical devices accessible to a Vulkan instance

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

fn enumerate_layer_properties() -> Result<Vec<VkLayerProperties>>

Returns up to all of global layer properties

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn enumerate_extension_properties(
layer_name: &str
) -> Result<Vec<VkExtensionProperties>>

Returns up to all of global extension properties

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_LAYER_NOT_PRESENT