Struct ferrite::MappedMemoryRange

pub struct MappedMemoryRange<'m>(_, _, _);

Specifies the block of mapped memory in a DeviceMemory

Methods

impl<'m> MappedMemoryRange<'m>

unsafe fn get<T>(&self, offset: usize) -> &T

Get a reference in mapped memory with byte offsets

Safety

Caller must guarantee that the pointer and its alignment are valid

unsafe fn get_mut<T>(&self, offset: usize) -> &mut T

Get a mutable reference in mapped memory with byte offsets

Safety

Caller must guarantee that the pointer and its alignment are valid

fn manual_flush(self) -> VkMappedMemoryRange

Flushes the memory range manually. Returns a structure for flush operation