Struct ferrite::PhysicalDevice

pub struct PhysicalDevice(_, _);

Opaque handle to a physical device object

Methods

impl PhysicalDevice

fn parent(&self) -> &Instance

fn features(&self) -> VkPhysicalDeviceFeatures

Reports capabilities of a physical device.

fn format_properties(&self, format: VkFormat) -> VkFormatProperties

Lists physical device's format capabilities

fn image_format_properties(
&self,
format: VkFormat,
itype: VkImageType,
tiling: VkImageTiling,
usage: ImageUsage,
flags: ImageFlags
) -> Result<VkImageFormatProperties>

Lists physical device's image format capabilities

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_FORMAT_NOT_SUPPORTED

fn properties(&self) -> VkPhysicalDeviceProperties

Returns properties of a physical device

fn queue_family_properties(&self) -> QueueFamilies

Reports properties of the queues of the specified physical device

fn memory_properties(&self) -> VkPhysicalDeviceMemoryProperties

Reports memory information for the specified physical device

fn sparse_image_format_properties(
&self,
format: VkFormat,
itype: VkImageType,
samples: VkSampleCountFlags,
usage: ImageUsage,
tiling: VkImageTiling
) -> Vec<VkSparseImageFormatProperties>

Retrieve properties of an image format applied to sparse images

fn surface_support(&self, queue_family: u32, surface: &Surface) -> Result<bool>

Query if presentation is supported

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

fn surface_capabilities(
&self,
surface: &Surface
) -> Result<VkSurfaceCapabilitiesKHR>

Query surface capabilities

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

fn surface_formats(&self, surface: &Surface) -> Result<Vec<VkSurfaceFormatKHR>>

Query color formats supported by surface

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

fn surface_present_modes(&self, surface: &Surface) -> Result<Vec<PresentMode>>

Query supported presentation modes

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_SURFACE_LOST_KHR

impl PhysicalDevice

VK_KHR_display

fn display_properties(&self) -> Result<Vec<VkDisplayPropertiesKHR>>

Query information about the available displays

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn display_plane_properties(&self) -> Result<Vec<VkDisplayPlanePropertiesKHR>>

Query the plane properties

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn display_plane_supported_displays(
&self,
index: u32
) -> Result<Vec<VkDisplayKHR>>

Query the list of displays a plane supports

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn display_mode_properties(
&self,
display: VkDisplayKHR
) -> Result<Vec<VkDisplayModePropertiesKHR>>

Query the set of mode properties supported by the display

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn new_display_mode(
&self,
display: VkDisplayKHR,
region: Extent2D,
refresh_rate: u32
) -> Result<VkDisplayModeKHR>

Create a display mode

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_INITIALIZATION_FAILED

fn display_plane_capabilities(
&self,
mode: VkDisplayModeKHR,
plane_index: u32
) -> Result<VkDisplayPlaneCapabilitiesKHR>

Query capabilities of a mode and plane combination

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY