Struct ferrite::PipelineStageFlags

pub struct PipelineStageFlags(pub VkPipelineStageFlags);

Bitmask specifying pipeline stages

Methods

impl PipelineStageFlags

const TOP_OF_PIPE: Self

TOP_OF_PIPE: Self = PipelineStageFlags(VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT)

The stage of the pipeline where any commands are initially received by the queue

const DRAW_INDIRECT: Self

DRAW_INDIRECT: Self = PipelineStageFlags(VK_PIPELINE_STAGE_DRAW_INDIRECT_BIT)

The stage of the pipeline where Draw/DispatchIndirect data structures are consumed

const VERTEX_INPUT: Self

VERTEX_INPUT: Self = PipelineStageFlags(VK_PIPELINE_STAGE_VERTEX_INPUT_BIT)

The stage of the pipeline where vertex and index buffers are consumed

const VERTEX_SHADER: Self

VERTEX_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_VERTEX_SHADER_BIT)

The vertex shader stage

const TESSELLATION_CONTROL_SHADER: Self

TESSELLATION_CONTROL_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_TESSELLATION_CONTROL_SHADER_BIT)

The tessellation control shader stage

const TESSELLATION_EVALUATION_SHADER: Self

TESSELLATION_EVALUATION_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_TESSELLATION_EVALUATION_SHADER_BIT)

The tessellation evaluation shader stage

const GEOMETRY_SHADER: Self

GEOMETRY_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_GEOMETRY_SHADER_BIT)

The geometry shader stage

const FRAGMENT_SHADER: Self

FRAGMENT_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_FRAGMENT_SHADER_BIT)

The fragment shader stage

const EARLY_FRAGMENT_TESTS: Self

EARLY_FRAGMENT_TESTS: Self = PipelineStageFlags(VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT)

The stage of the pipeline where early fragment tests (depth and stencil tests before fragment shading) are performed

const LATE_FRAGMENT_TESTS: Self

LATE_FRAGMENT_TESTS: Self = PipelineStageFlags(VK_PIPELINE_STAGE_LATE_FRAGMENT_TESTS_BIT)

The stage of the pipeline where late fragment tests (depth and stencil tests after fragment shading) are performed

const COLOR_ATTACHMENT_OUTPUT: Self

COLOR_ATTACHMENT_OUTPUT: Self = PipelineStageFlags(VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT)

The stage of the pipeline after blending where the final color values are output from the pipeline

const TRANSFER: Self

TRANSFER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_TRANSFER_BIT)

The execution of copy commands

const COMPUTE_SHADER: Self

COMPUTE_SHADER: Self = PipelineStageFlags(VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT)

The execution of a compute shader

const BOTTOM_OF_PIPE: Self

BOTTOM_OF_PIPE: Self = PipelineStageFlags(VK_PIPELINE_STAGE_BOTTOM_OF_PIPE_BIT)

The final stage in the pipeline where operations generated by all commands complete execution

const HOST: Self

HOST: Self = PipelineStageFlags(VK_PIPELINE_STAGE_HOST_BIT)

A pseudo-stage indicating execution on the host of reads/writes of device memory

const ALL_GRAPHICS: Self

ALL_GRAPHICS: Self = PipelineStageFlags(VK_PIPELINE_STAGE_ALL_GRAPHICS_BIT)

The execution of all graphics pipeline stages

const ALL_COMMANDS: Self

ALL_COMMANDS: Self = PipelineStageFlags(VK_PIPELINE_STAGE_ALL_COMMANDS_BIT)

Equivalent to the logical OR of every other pipeline stage flag that is supported on the quue it is used with

fn top_of_pipe(&self) -> Self

The stage of the pipeline where any commands are initially received by the queue

fn draw_indirect(&self) -> Self

The stage of the pipeline where Draw/DispatchIndirect data structures are consumed

fn vertex_input(&self) -> Self

The stage of the pipeline where vertex and index buffers are consumed

fn vertex_shader(&self) -> Self

The vertex shader stage

fn tessellation_control_shader(&self) -> Self

The tessellation control shader stage

fn tessellation_evaluation_shader(&self) -> Self

The tessellation evaluation shader stage

fn geometry_shader(&self) -> Self

The geometry shader stage

fn fragment_shader(&self) -> Self

The fragment shader stage

fn early_fragment_tests(&self) -> Self

The stage of the pipeline where early fragment tests (depth and stencil tests before fragment shading) are performed

fn late_fragment_tests(&self) -> Self

The stage of the pipeline where late fragment tests (depth and stencil tests after fragment shading) are performed

fn color_attachment_output(&self) -> Self

The stage of the pipeline after blending where the final color values are output from the pipeline

fn transfer(&self) -> Self

The execution of copy commands

fn compute_shader(&self) -> Self

The execution of a compute shader

fn bottom_of_pipe(&self) -> Self

The final stage in the pipeline where operations generated by all commands complete execution

fn host(&self) -> Self

A pseudo-stage indicating execution on the host of reads/writes of device memory

fn all_graphics(&self) -> Self

The execution of all graphics pipeline stages

fn all_commands(&self) -> Self

Equivalent to the logical OR of every other pipeline stage flag that is supported on the quue it is used with