Struct ferrite::QueryPipelineStatisticFlags

pub struct QueryPipelineStatisticFlags(pub VkQueryPipelineStatisticFlags);

Bitmask specifying queried pipeline statistics

Methods

impl QueryPipelineStatisticFlags

const INPUT_ASSEMBLY_VERTICES: Self

INPUT_ASSEMBLY_VERTICES: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_VERTICES_BIT)

Queries managed by the pool will count the number of vertices processed by the input assembly stage

const INPUT_ASSEMBLY_PRIMITIVES: Self

INPUT_ASSEMBLY_PRIMITIVES: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_INPUT_ASSEMBLY_PRIMITIVES_BIT)

Queries managed by the pool will count the number of primitives processed by the input assembly state

const VERTEX_SHADER_INVOCATIONS: Self

VERTEX_SHADER_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_VERTEX_SHADER_INVOCATIONS_BIT)

Queries managed by the pool will count the number of vertex shader invocations

const GEOMETRY_SHADER_INVOCATIONS: Self

GEOMETRY_SHADER_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_INVOCATIONS_BIT)

Queries managed by the pool will count the number of geometry shader invocations

const GEOMETRY_SHADER_PRIMITIVES: Self

GEOMETRY_SHADER_PRIMITIVES: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_GEOMETRY_SHADER_PRIMITIVES_BIT)

Queries managed by the pool will count the number of primitives generated by geometry shader invocations

const CLIPPING_INVOCATIONS: Self

CLIPPING_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_CLIPPING_INVOCATIONS_BIT)

Queries managed by the pool will count the number of primitives processed by the Primitive Clipping stage of the pipeline

const CLIPPING_PRIMITIVES: Self

CLIPPING_PRIMITIVES: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_CLIPPING_PRIMITIVES_BIT)

Queries managed by the pool will count the number of primitives output by the Primitive Clipping stage of the pipeline

const FRAGMENT_SHADER_INVOCATIONS: Self

FRAGMENT_SHADER_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_FRAGMENT_SHADER_INVOCATIONS_BIT)

Queries managed by the pool will count the number of fragment shader invocations

const TESSELLATION_CONTROL_SHADER_PATCHES: Self

TESSELLATION_CONTROL_SHADER_PATCHES: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_TESSELLATION_CONTROL_SHADER_PATCHES_BIT)

Queries managed by the pool will count the number of patches processed by the tessellation control shader

const TESSELLATION_EVALUATION_SHADER_INVOCATIONS: Self

TESSELLATION_EVALUATION_SHADER_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_TESSELLATION_EVALUATION_SHADER_INVOCATIONS_BIT)

Queries managed by the pool will count the number of invocations of the tessellation evaluation shader

const COMPUTE_SHADER_INVOCATIONS: Self

COMPUTE_SHADER_INVOCATIONS: Self = QueryPipelineStatisticFlags(VK_QUERY_PIPELINE_STATISTIC_COMPUTE_SHADER_INVOCATIONS_BIT)

Queries managed by the pool will count the number of compute shader invocations

fn input_assembly_vertices(&self) -> Self

Queries managed by the pool will count the number of vertices processed by the input assembly stage

fn input_assembly_primitives(&self) -> Self

Queries managed by the pool will count the number of primitives processed by the input assembly state

fn vertex_shader_invocations(&self) -> Self

Queries managed by the pool will count the number of vertex shader invocations

fn geometry_shader_invocations(&self) -> Self

Queries managed by the pool will count the number of geometry shader invocations

fn geometry_shader_primitives(&self) -> Self

Queries managed by the pool will count the number of primitives generated by geometry shader invocations

fn clipping_invocations(&self) -> Self

Queries managed by the pool will count the number of primitives processed by the Primitive Clipping stage of the pipeline

fn clipping_primitives(&self) -> Self

Queries managed by the pool will count the number of primitives output by the Primitive Clipping stage of the pipeline

fn fragment_shader_invocations(&self) -> Self

Queries managed by the pool will count the number of fragment shader invocations

fn tessellation_control_shader_patches(&self) -> Self

Queries managed by the pool will count the number of patches processed by the tessellation control shader

fn tessellation_evaluation_shader_invocations(&self) -> Self

Queries managed by the pool will count the number of invocations of the tessellation evaluation shader

fn compute_shader_invocations(&self) -> Self

Queries managed by the pool will count the number of compute shader invocations