Struct ferrite::QueryPool

pub struct QueryPool(_, _);

Opaque handle to a query pool object

Methods

impl QueryPool

fn new(device: &Device, qtype: QueryType, count: u32) -> Result<Self>

Create a new query pool object

Failure

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

fn results64(
&self,
query_range: Range<u32>,
flags: QueryResultFlags
) -> Result<Vec<u64>>

Copy results of queries in a query pool to a host memory region

Failure

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

fn results32(
&self,
query_range: Range<u32>,
flags: QueryResultFlags
) -> Result<Vec<u32>>

Copy results of queries in a query pool to a host memory region

Failure

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST