Struct ferrite::QueueFamilies

pub struct QueueFamilies(pub Vec<VkQueueFamilyProperties>);

List of queue families

Methods

impl QueueFamilies

fn find_matching_index(&self, flags: QueueFlags) -> Option<u32>

Find a queue family index containing specified bitflags

fn find_another_matching_index(
&self,
flags: QueueFlags,
exclude: u32
) -> Option<u32>

Find a queue family index containing specified bitflags

fn count(&self) -> u32

Number of queue families

fn queue_count(&self, family_index: u32) -> u32

Number of queues in selected queue family

fn timestamp_valid_bits(&self, family_index: u32) -> u32

Unsigned integer count of meaningful bits in the timestamps written via vkCmdWriteTimestamp

fn minimum_image_transfer_granularity(&self, family_index: u32) -> &VkExtent3D

Minimum granularity supported for image transfer operations on the queues in selected queue family