Struct ferrite::Swapchain

pub struct Swapchain(_);

Opaque handle to a swapchain object

Methods

impl Swapchain

Following methods are enabled with [feature = "FeImplements" and feature = "VK_KHR_swapchain"]

pub fn get_images(&self) -> Result<Vec<Image>>

Obtain the array of presentable images associated with a swapchain

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

impl Swapchain

pub fn format(&self) -> VkFormat

pub fn size(&self) -> &Extent3D

impl Swapchain

pub fn acquire_next(
&self,
timeout: Option<u64>,
completion: CompletionHandler
) -> Result<u32>

Retrieve the index of the next available presentation image

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR

pub fn queue_present(
&self,
queue: &Queue,
index: u32,
wait_semaphores: &[&Semaphore]
) -> Result<()>

Queue an image for presentation

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_OUT_OF_DATE_KHR

VK_ERROR_SURFACE_LOST_KHR