Struct ferrite::SwapchainBuilder

pub struct SwapchainBuilder<'d>(_, _, _);

Builder object to construct a Swapchain

Methods

impl<'d> SwapchainBuilder<'d>

pub fn new(
surface: &'d Surface,
min_image_count: u32,
format: VkSurfaceFormatKHR,
extent: Extent2D,
usage: ImageUsage
) -> Self

pub fn array_layers(&mut self, layers: u32) -> &mut Self

pub fn share(&mut self, queue_families: Vec<u32>) -> &mut Self

pub fn pre_transform(&mut self, tf: SurfaceTransform) -> &mut Self

pub fn composite_alpha(&mut self, a: CompositeAlpha) -> &mut Self

pub fn present_mode(&mut self, mode: PresentMode) -> &mut Self

pub fn enable_clip(&mut self) -> &mut Self

Enables whether the Vulkan implementation is allowed to discard rendering operations that affect regions of the surface which aren't visible

pub fn create(&self, device: &Device) -> Result<Swapchain>

Create a swapchain

Failures

On failure, this command returns

VK_ERROR_OUT_OF_HOST_MEMORY

VK_ERROR_OUT_OF_DEVICE_MEMORY

VK_ERROR_DEVICE_LOST

VK_ERROR_SURFACE_LOST_KHR

VK_ERROR_NATIVE_WINDOW_IN_USE_KHR